Move command to more logical place.
